EDM-1695: Add middleware to set org id in api

+++ internal/service/certificatesigningrequest.go
@@ -12,6 +12,7 @@ import (
 	"github.com/flightctl/flightctl/internal/crypto"
 	"github.com/flightctl/flightctl/internal/flterrors"
 	"github.com/flightctl/flightctl/internal/store/selector"
+	"github.com/flightctl/flightctl/internal/util"
 	"github.com/google/uuid"
 )
 
@@ -104,7 +105,10 @@ func signApprovedCertificateSigningRequest(ca *crypto.CAClient, request api.Cert
 }
 
 func (h *ServiceHandler) ListCertificateSigningRequests(ctx context.Context, params api.ListCertificateSigningRequestsParams) (*api.CertificateSigningRequestList, api.Status) {
-	orgId := GetOrgIdFromContext(ctx)
+	orgId, ok := util.GetOrgIdFromContext(ctx)
+	if !ok {
+		return nil, api.StatusBadRequest(flterrors.ErrInvalidOrganizationID.Error())
+	}
 
 	listParams, status := prepareListParams(params.Continue, params.LabelSelector, params.FieldSelector, params.Limit)
 	if status != api.StatusOK() {
@@ -148,7 +152,10 @@ func (h *ServiceHandler) verifyCSRParameters(ctx context.Context, csr api.Certif
 }
 
 func (h *ServiceHandler) CreateCertificateSigningRequest(ctx context.Context, csr api.CertificateSigningRequest) (*api.CertificateSigningRequest, api.Status) {
-	orgId := GetOrgIdFromContext(ctx)
+	orgId, ok := util.GetOrgIdFromContext(ctx)
+	if !ok {
+		return nil, api.StatusBadRequest(flterrors.ErrInvalidOrganizationID.Error())
+	}
 
 	// don't set fields that are managed by the service
 	csr.Status = nil
@@ -183,7 +190,10 @@ func (h *ServiceHandler) CreateCertificateSigningRequest(ctx context.Context, cs
 }
 
 func (h *ServiceHandler) DeleteCertificateSigningRequest(ctx context.Context, name string) api.Status {
-	orgId := GetOrgIdFromContext(ctx)
+	orgId, ok := util.GetOrgIdFromContext(ctx)
+	if !ok {
+		return api.StatusBadRequest(flterrors.ErrInvalidOrganizationID.Error())
+	}
 
 	deleted, err := h.store.CertificateSigningRequest().Delete(ctx, orgId, name)
 	status := StoreErrorToApiStatus(err, false, api.CertificateSigningRequestKind, &name)
@@ -194,14 +204,20 @@ func (h *ServiceHandler) DeleteCertificateSigningRequest(ctx context.Context, na
 }
 
 func (h *ServiceHandler) GetCertificateSigningRequest(ctx context.Context, name string) (*api.CertificateSigningRequest, api.Status) {
-	orgId := GetOrgIdFromContext(ctx)
+	orgId, ok := util.GetOrgIdFromContext(ctx)
+	if !ok {
+		return nil, api.StatusBadRequest(flterrors.ErrInvalidOrganizationID.Error())
+	}
 
 	result, err := h.store.CertificateSigningRequest().Get(ctx, orgId, name)
 	return result, StoreErrorToApiStatus(err, false, api.CertificateSigningRequestKind, &name)
 }
 
 func (h *ServiceHandler) PatchCertificateSigningRequest(ctx context.Context, name string, patch api.PatchRequest) (*api.CertificateSigningRequest, api.Status) {
-	orgId := GetOrgIdFromContext(ctx)
+	orgId, ok := util.GetOrgIdFromContext(ctx)
+	if !ok {
+		return nil, api.StatusBadRequest(flterrors.ErrInvalidOrganizationID.Error())
+	}
 
 	currentObj, err := h.store.CertificateSigningRequest().Get(ctx, orgId, name)
 	if err != nil {
@@ -249,7 +265,10 @@ func (h *ServiceHandler) PatchCertificateSigningRequest(ctx context.Context, nam
 }
 
 func (h *ServiceHandler) ReplaceCertificateSigningRequest(ctx context.Context, name string, csr api.CertificateSigningRequest) (*api.CertificateSigningRequest, api.Status) {
-	orgId := GetOrgIdFromContext(ctx)
+	orgId, ok := util.GetOrgIdFromContext(ctx)
+	if !ok {
+		return nil, api.StatusBadRequest(flterrors.ErrInvalidOrganizationID.Error())
+	}
 
 	// don't overwrite fields that are managed by the service
 	csr.Status = nil
@@ -286,7 +305,10 @@ func (h *ServiceHandler) ReplaceCertificateSigningRequest(ctx context.Context, n
 
 // NOTE: Approval currently also issues a certificate - this will change in the future based on policy
 func (h *ServiceHandler) UpdateCertificateSigningRequestApproval(ctx context.Context, name string, csr api.CertificateSigningRequest) (*api.CertificateSigningRequest, api.Status) {
-	orgId := GetOrgIdFromContext(ctx)
+	orgId, ok := util.GetOrgIdFromContext(ctx)
+	if !ok {
+		return nil, api.StatusBadRequest(flterrors.ErrInvalidOrganizationID.Error())
+	}
 
 	newCSR := &csr
 	NilOutManagedObjectMetaProperties(&newCSR.Metadata)
